+++ driver.go
@@ -370,6 +370,7 @@ func (d *Driver) StartTask(cfg *drivers.TaskConfig) (*drivers.TaskHandle, *drive
 	createOpts.ContainerBasicConfig.Hostname = driverConfig.Hostname
 	createOpts.ContainerBasicConfig.Sysctl = driverConfig.Sysctl
 	createOpts.ContainerBasicConfig.Terminal = driverConfig.Tty
+	createOpts.ContainerBasicConfig.Labels = driverConfig.Labels
 
 	createOpts.ContainerBasicConfig.LogConfiguration.Path = cfg.StdoutPath