+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * config: Added tty option
 * config: Support for sysctl configuration [[GH-82](https://github.com/hashicorp/nomad-driver-podman/issues/82)]
 * config: Fixed a bug where we always pulled an image if image name has a transport prefix [[GH-88](https://github.com/hashicorp/nomad-driver-podman/pull/88)]
+* config: Added labels option
 * config: Add force_pull option
 
 BUG FIXES: